Refactor(close #5): 예외 처리 추가 및 삭제 시 userid 확인

+++ src/main/java/efub/lead/twitter/service/TweetService.java
@@ -12,6 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -78,8 +79,12 @@ public class TweetService {
     }
 
     @Transactional
-    public void deleteTweet(Long tweetId){
-        tweetRepository.deleteById(tweetId);
+    public void deleteTweet(Long tweetId, Long userId){
+        User user = tweetRepository.findById(tweetId).get().getUser();
+        if (Objects.equals(user.getUserId(), userId))
+            tweetRepository.deleteById(tweetId);
+        else
+            throw new IllegalStateException();
     }
 
     public TweetResponseDTO buildTweetDTO(Tweet tweet) {
